Fix test and remove PushStatusStoreReocrdDeleter
