+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/PushStatusStoreTest.java
@@ -42,7 +42,7 @@ import com.linkedin.venice.pubsub.PubSubTopicRepository;
 import com.linkedin.venice.pubsub.api.PubSubTopic;
 import com.linkedin.venice.pushmonitor.ExecutionStatus;
 import com.linkedin.venice.pushstatushelper.PushStatusStoreReader;
-import com.linkedin.venice.pushstatushelper.PushStatusStoreRecordDeleter;
+import com.linkedin.venice.pushstatushelper.PushStatusStoreWriter;
 import com.linkedin.venice.schema.writecompute.WriteComputeSchemaConverter;
 import com.linkedin.venice.serialization.avro.AvroProtocolDefinition;
 import com.linkedin.venice.utils.DataProviderUtils;
@@ -266,15 +266,17 @@ public class PushStatusStoreTest {
         response = controllerClient.queryJobStatus(job.getTopicToMonitor(), job.getIncrementalPushVersion());
         assertEquals(response.getStatus(), ExecutionStatus.END_OF_INCREMENTAL_PUSH_RECEIVED.name());
 
-        PushStatusStoreRecordDeleter statusStoreDeleter = new PushStatusStoreRecordDeleter(
+        PushStatusStoreWriter statusStoreWriter = new PushStatusStoreWriter(
             cluster.getLeaderVeniceController().getVeniceHelixAdmin().getVeniceWriterFactory(),
+            "dummyInstance",
+            1,
             WriteComputeSchemaConverter.getInstance()
                 .convertFromValueRecordSchema(
                     AvroProtocolDefinition.PUSH_STATUS_SYSTEM_SCHEMA_STORE.getCurrentProtocolVersionSchema()));
 
         // After deleting the inc push status belonging to just one partition we should expect
         // SOIP from the controller since other partition has replicas with EOIP status
-        statusStoreDeleter.deletePartitionIncrementalPushStatus(storeName, 1, incPushVersion.get(), 1).get();
+        statusStoreWriter.deletePartitionIncrementalPushStatus(storeName, 1, incPushVersion.get(), 1).get();
         TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
           // N.B.: Even though we block on the deleter's future, that only means the delete message is persisted into
           // Kafka, but querying the system store may still yield a stale result, hence the need for retrying.
@@ -284,7 +286,7 @@ public class PushStatusStoreTest {
         });
 
         // expect NOT_CREATED when statuses of all partitions are not available in the push status store
-        statusStoreDeleter.deletePartitionIncrementalPushStatus(storeName, 1, incPushVersion.get(), 0).get();
+        statusStoreWriter.deletePartitionIncrementalPushStatus(storeName, 1, incPushVersion.get(), 0).get();
         TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
           JobStatusQueryResponse jobStatusQueryResponse =
               controllerClient.queryJobStatus(job.getTopicToMonitor(), job.getIncrementalPushVersion());
