+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/Admin.java
@@ -28,7 +28,6 @@ import com.linkedin.venice.persona.StoragePersona;
 import com.linkedin.venice.pubsub.PubSubConsumerAdapterFactory;
 import com.linkedin.venice.pushmonitor.ExecutionStatus;
 import com.linkedin.venice.pushstatushelper.PushStatusStoreReader;
-import com.linkedin.venice.pushstatushelper.PushStatusStoreRecordDeleter;
 import com.linkedin.venice.pushstatushelper.PushStatusStoreWriter;
 import com.linkedin.venice.schema.GeneratedSchemaID;
 import com.linkedin.venice.schema.SchemaEntry;
@@ -748,11 +747,6 @@ public interface Admin extends AutoCloseable, Closeable {
   MetaStoreWriter getMetaStoreWriter();
 
   MetaStoreReader getMetaStoreReader();
-
-  /**
-   * Return {@link PushStatusStoreRecordDeleter}.
-   */
-  PushStatusStoreRecordDeleter getPushStatusStoreRecordDeleter();
 
   /** Get a list of clusters this controller is a leader of.
    * @return a list of clusters this controller is a leader of.
