+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/VeniceParentHelixAdmin.java
@@ -188,7 +188,6 @@ import com.linkedin.venice.pubsub.api.PubSubProduceResult;
 import com.linkedin.venice.pubsub.api.PubSubTopic;
 import com.linkedin.venice.pushmonitor.ExecutionStatus;
 import com.linkedin.venice.pushstatushelper.PushStatusStoreReader;
-import com.linkedin.venice.pushstatushelper.PushStatusStoreRecordDeleter;
 import com.linkedin.venice.pushstatushelper.PushStatusStoreWriter;
 import com.linkedin.venice.schema.AvroSchemaParseUtils;
 import com.linkedin.venice.schema.GeneratedSchemaID;
@@ -4758,14 +4757,6 @@ public class VeniceParentHelixAdmin implements Admin {
   }
 
   /**
-   * @see Admin#getPushStatusStoreRecordDeleter()
-   */
-  @Override
-  public PushStatusStoreRecordDeleter getPushStatusStoreRecordDeleter() {
-    return getVeniceHelixAdmin().getPushStatusStoreRecordDeleter();
-  }
-
-  /**
    * @see Admin#getEmergencySourceRegion(String)
    */
   @Override
@@ -5241,7 +5232,7 @@ public class VeniceParentHelixAdmin implements Admin {
 
   @Override
   public PushStatusStoreWriter getPushStatusStoreWriter() {
-    throw new VeniceUnsupportedOperationException("Parent controller does not have Da Vinci push status store writer");
+    return getVeniceHelixAdmin().getPushStatusStoreWriter();
   }
 
   @Override
